requirements/jenkins.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/jenkins.txt requirements/base.txt requirements/dev.txt requirements/jenkins.in
#
astroid==2.3.1
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cssselect==1.1.0
django-appconf==1.0.3
django-axes==5.0.17
django-choices==1.7.0
django-cors-middleware==1.4.0
django-debug-toolbar==2.0
django-extensions==2.2.3
django-filter==2.2.0
django-ipware==2.1.0
django-jenkins==0.110.0
django-markup==1.4.1
django-rest-framework-condition==0.1.1
django-solo==1.1.3
django-webtest==1.9.7
django==2.2.6
djangorestframework-camel-case==1.1.1
djangorestframework==3.9.4
drf-nested-routers==0.91
drf-yasg==1.16.0
factory-boy==2.12.0
faker==2.0.3
freezegun==0.3.12
gemma-zds-client==0.13.0
idna==2.8
inflection==0.3.1
iso-639==0.4.5
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jinja2==2.10.3
lazy-object-proxy==1.4.2
lxml==4.4.1
markdown==3.1.1
markupsafe==1.1.1
mccabe==0.6.1
mock==3.0.5
oyaml==0.9
pep8==1.7.1
pip-tools==4.2.0
psycopg2==2.8.3
pyjwt==1.7.1
pylint==2.4.2
pyquery==1.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
pyyaml==5.1.2
raven==6.10.0
requests==2.22.0
ruamel.yaml.clib==0.2.0
ruamel.yaml==0.16.5
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tblib==1.4.0
text-unidecode==1.3
typed-ast==1.4.0
unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.6
uwsgi==2.0.18
vng-api-common==1.3.2
waitress==1.3.1
webob==1.8.5
webtest==2.0.33
wrapt==1.11.2

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.4.0        # via markdown
